+++ scripts/install/nginx.sh
@@ -57,7 +57,3 @@ fi
 
 # Logs
 mkdir -p $DATA_DIRECTORY/logs/nginx
-
-# Loading permissions
-chown -R www-data:www-data \
-    $DATA_DIRECTORY/logs \