+++ scripts/install/php.sh
@@ -21,7 +21,3 @@ mkdir -p $DATA_DIRECTORY/logs/php-fpm
 mkdir -p $DATA_DIRECTORY/logs/php-cli
 sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-fpm/errors.log|g" /etc/php5/fpm/php.ini
 sed -i "s|;error_log = php_errors.log|error_log = ${DATA_DIRECTORY}/logs/php-cli/errors.log|g" /etc/php5/cli/php.ini
-
-# Loading permissions
-chown -R www-data:www-data \
-    $DATA_DIRECTORY/logs \