+++ scripts/launch.sh
@@ -9,12 +9,17 @@ SCRIPTS_DIRECTORY=/scripts/toran-proxy
 if [ -d $DATA_DIRECTORY/logs ]; then
     rm -rf $DATA_DIRECTORY/logs
 fi
+mkdir $DATA_DIRECTORY/logs
 
 # Initilisation
 source $SCRIPTS_DIRECTORY/install/php.sh
 source $SCRIPTS_DIRECTORY/install/nginx.sh
 source $SCRIPTS_DIRECTORY/install/ssh.sh
 source $SCRIPTS_DIRECTORY/install/toran.sh
+
+# Loading logs permissions
+chown -R www-data:www-data \
+    $DATA_DIRECTORY/logs \
 
 # Start services
 echo "Starting Toran Proxy..."