+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 **latest**
+- Upgrade toran proxy to version 1.5.2
 - Redirection `http` to `https` auto with use https vhost type
 - Generate self-signed certificates automatically
 - Clean create and permissions on logs directory